Sensitivity cleanup: top-level imports, merge report CLI, eval/ outputs
- move all in-function imports (sbi, matplotlib, scipy) to module top; these
  analysis modules already import torch at top and are only loaded when running
  an analysis, so lazy imports added no value
- merge report.py into generate_report.py and delete report.py
- write generated reports to eval/ (gitignored) instead of the repo root; stop
  tracking the sensitivity_*.png demo artifacts committed by mistake
- drop 'robolab' mentions and remaining backticks from sensitivity docstrings
- plot_marginals docstring: drop the 'whole deliverable' sentence
