+++ .gitignore
@@ -1,5 +1,7 @@
 **/build
 
+# Eval outputs (generated sensitivity reports, etc.)
+/eval/
 
 # Prerequisites
 *.d
