--- sensitivity_mixed.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:52c02b579b63f11ca5fa1bcd009840eef8022faec7e0a87575cb63b1ad6d2a8d
-size 96934
