--- sensitivity_rich.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:314e08938b0043b939df50c94a45ceac71c8f182ee4de120fa13da9194719fb7
-size 189814
