--- sensitivity_synthetic.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9ea649f10a53d11a4069154824190bbabc57e1697e1755c5fe87dba6393620c6
-size 77273
